Electric Boogaloo No1
Roberto should infiltrate to the Electric Boogaloo, a bar
managed by the local mafia Villa Viciosa in order to solve the murder of
his wife. To achieve his mission, shall be prepared in
dark arts of "Palo mango", dancing tasty and reach to the end of
salsa contest in the bar that names this comic.

Roberto López will have to infiltrate in Electric Boogaloo, a bar
managed by the mob from ‘Villa Viciosa’, so he can solve the murder of
his ex-wife. To achieve his mission, he will prepare in the dark art of
‘Palo de Mango’, becoming the best salsa dancer in town.